requirements.txt
# Web Framework
streamlit~=1.35.0

# LangChain Core & Integrations
langchain~=0.2.1
langchain-google-genai~=1.0.4
langchain-community~=0.2.1

# Vector Database
chromadb~=0.5.0

# PDF Parsing
pypdf~=4.2.0

# Environment Variables
python-dotenv~=1.0.1

# Google AI SDK (implicitly required by langchain-google-genai but good to specify)
google-generativeai~=0.7.0

# Exam Export Libraries
reportlab~=4.2.0
python-docx~=1.1.2 